+++ src/auth/auth.module.ts
@@ -3,10 +3,9 @@ import { ConfigModule } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
 import { MongooseModule } from '@nestjs/mongoose';
 import { PassportModule } from '@nestjs/passport';
-
 import { UsersSchema } from 'src/users/schemas/users.schema';
 import { AuthService } from './auth.service';
-import { AuthController } from './controllers/auth.controller';
+import { AuthController } from './auth.controller';
 import { JwtStrategy } from '../common/strategies/jwt.strategy';
 import { LocalStrategy } from './strategies/local.strategy';
 import { UsersService } from 'src/users/users.service';
