+++ src/users/users.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 
-import { UsersController } from './controllers/users.controller';
+import { UsersController } from './users.controller';
 import { UsersSchema } from './schemas/users.schema';
 import { UsersService } from './users.service';
